Started work on a hover module
There's now a simple system to bind a list of arbitrary functions to be
called whenever an svg element that was rendered by soda is hovered.
Functions can be written that expect a reference to a d3 selection to
the hovered svg and to the Annotation object that was bound to that svg.