Deterministic graph color layout, graph object
I implemented a deterministic graph coloring algorithm that simply makes
a greedy coloring by taking as many of the largest unplaced annotations
and placing them in a layer. It honestly seems to be just as optimal as
the heuristic.

In other news, I added an AnnotationGraph object since I'm going to plan
to make a few other graph based layout algorithms.