Resize now works without a zoom controller
Now, TrackCharts have the capability to resize themselves without the
use of a ZoomController. To accomodate this, I made a substantial change
to the way rendering is done on charts. Now, the ChartBase class has a
default render() implementation, which saves the last used render
parameters. This makes it possible to arbitrarily re-render the chart
without recieving new data.