Started to work on GFF3 parser
I made a simple Gff3Annotation class, which will probably be messed with
quite a bit before I'm done. To accomodate GFF3 data, I added the
'Unknown' and 'Unoriented' types to the Orientation enum. I wrote a
simple parser for GFF3, which is untested as of now. I'm going to need
to think carefully about how I'm going to resolve children/parent
records in GFF3 data.