Completely redid the GFF3 parser
So, it turns out the GFF3 spec is horrifying. So, I'm using an npm
package that seems to be created for JBrowse called gff-js, which
handles the parsing. It seems to work pretty well, but it is written in
JavaScript, so there's some nasty bandaid fixes to make it work. In
particular, I had to provide an 'empty definition' (src/decs.d.ts) for
the library. This basically just tells TypeScript that the module does
indeed exist, and it defaults all types that are spit out of the library
to be defined as 'any.'

I seem to be getting the right number of annotation objects, now I just
need to mess around with trying to visualize them.