Rewrote all the GFF3 parsing...
The GFF3 format continually reveals itself to be more and more convoluted. At this point, I've got a parser that creates a SODA object representation for GFF3 records that satisfies me, and I'm pretty close to having a post pprocessing step figured out to group records with the pattern described as the "canonical gene representation" (https://github.com/The-Sequence-Ontology/Specifications/blob/master/gff3.md#the-canonical-gene). It seems like the official canonical convention is the one we should explicitly support.