Added BufferedQueryParameters, fixed layout bug
BufferedQueryParameters can be used to load more data than is requested
by a query so that the visualization can be buffered, preventing
immediate re-querying on slight panning/zooming. The QueryController
will check whether or not a query it was given is buffered.

The layout algorithms now don't error out when given empty lists of
Annotations to work with.