Started to add brush module
The brush module works a lot like the click and hover modules. Brush
behaviors can be added to charts and execute callback functions.

It turns out that d3 brushing and zooming really don't get along, so I
patched around this in the TrackChart. I'll need to remember to fix this
later, which may end up becoming a full disentanglement of zooming from
the TrackChart implementation.