Preparing to work on sequence glyphs
I'm branching this off of the brush branch, since it's convenient for my
pratical testing of the brush module to have sequence glyphs working
better.

I'm tweaking the sequence glyph module so that it hopefully can be a bit
more useful. I also added css IDs to the other glyphs, which will
probably be useful later.